Move LTEEntity to component
