+++ homeassistant/components/netgear_lte/sensor.py
@@ -4,8 +4,7 @@ import logging
 from homeassistant.components.sensor import DOMAIN
 from homeassistant.exceptions import PlatformNotReady
 
-from . import CONF_MONITORED_CONDITIONS, DATA_KEY
-from .lte_entity import LTEEntity
+from . import CONF_MONITORED_CONDITIONS, DATA_KEY, LTEEntity
 from .sensor_types import SENSOR_SMS, SENSOR_USAGE, SENSOR_UNITS
 
 DEPENDENCIES = ['netgear_lte']
